backend_py/pyproject.toml
[project]
name = "qa-dashboard-backend-py"
version = "3.0.0"
description = "QA Dashboard backend — FastAPI + SQLAlchemy rewrite"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "aiosqlite>=0.20",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.17",
    "httpx>=0.27",
    "tenacity>=9.0",
    "cachetools>=5.5",
    "prometheus-client>=0.21",
    "sentry-sdk[fastapi]>=2.19",
    "aiosmtplib>=3.0",
    "jinja2>=3.1",
    "openpyxl>=3.1",
    "python-pptx>=1.0",
    "playwright>=1.48",
    "boto3>=1.35",
    "aiobotocore>=2.15",
    "apscheduler>=3.10",
    "slowapi>=0.1.9",
    "structlog>=24.4",
    "orjson>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-passlib>=1.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
